+++ lib/model/Product.g.dart
@@ -22,6 +22,7 @@ Product _$ProductFromJson(Map<String, dynamic> json) {
     categories: json['categories'] as String,
     nutrimentEnergyUnit: json['nutriment_energy_unit'] as String,
     nutrimentDataPer: json['nutrition_data_per'] as String,
+    nutriscore: json['nutrition_grade_fr'] as String,
     nutriments: json['nutriments'] == null
         ? null
         : Nutriments.fromJson(json['nutriments'] as Map<String, dynamic>),
@@ -30,6 +31,10 @@ Product _$ProductFromJson(Map<String, dynamic> json) {
         : Additives.fromJson((json['additives_tags'] as List)
             ?.map((e) => e as String)
             ?.toList()),
+    nutrientLevels: json['nutrient_levels'] == null
+        ? null
+        : NutrientLevels.fromJson(
+            json['nutrient_levels'] as Map<String, dynamic>),
   )
     ..productNameFR = json['product_name_fr'] as String
     ..selectedImages = json['selected_images'] == null
@@ -74,12 +79,14 @@ Map<String, dynamic> _$ProductToJson(Product instance) {
   writeNotNull('ingredients', instance.ingredients);
   writeNotNull('nutriments', instance.nutriments);
   writeNotNull('additives_tags', instance.additives);
+  writeNotNull('nutrient_levels', instance.nutrientLevels);
   writeNotNull('ingredients_text', instance.ingredientsText);
   writeNotNull('ingredients_text_de', instance.ingredientsTextDE);
   writeNotNull('ingredients_text_en', instance.ingredientsTextEN);
   writeNotNull('ingredients_text_fr', instance.ingredientsTextFR);
   writeNotNull('nutriment_energy_unit', instance.nutrimentEnergyUnit);
   writeNotNull('nutrition_data_per', instance.nutrimentDataPer);
+  writeNotNull('nutrition_grade_fr', instance.nutriscore);
   writeNotNull('categories', instance.categories);
   writeNotNull('categories_tags', instance.categoriesTags);
   writeNotNull('labels_tags', instance.labelsTags);