+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [0.1.12] - 18.11.2019
+- nutrient levels added
+
+## [0.1.11] - 9.10.2019
+- additives management implemented
+- added Nutriscore to product
+- bugfix: ingredientText language dependent
+
 ## [0.1.10] - 5.10.2019
 - nutriments management implemented